+++ core/src/main/kotlin/service/IcebergGraphBuilder.kt
@@ -67,6 +67,7 @@ object IcebergGraphBuilder {
             unreferencedFiles = DeferredRead.of { findUnreferencedFiles(tableModel) },
             missingFiles = DeferredRead.of { tableModel.findMissingFiles() },
             expiryFiles = DeferredRead.of { tableModel.expiryFileInput() },
+            rewriteTablePath = DeferredRead.of { tableModel.rewriteTablePathInput() },
             integrity = DeferredRead.of { tableModel.integrityReport() },
             fileStats = DeferredRead.of { tableModel.fileStatsTargets() },
             statisticsFiles = newestMetadata?.let { meta ->